backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-groq
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
pydantic
sqlalchemy
sqlalchemy-utils
passlib[bcrypt]
python-jose[cryptography]
pandas
psycopg2-binary
requests
fpdf2
